[maint] Allow rpmbuild to work with Cygwin
Recent packaging changes related to TLS work lost changes that
allowed the DNE PAQX RPM to be built from Windows under
Cygwin. This change restores that ability. The following changes
were required:

- The `LOCAL_WORKSPACE` path variable for
  `dne-paqx-packaging/target` is calculated in `build_rpm.sh`
  rather than set in the maven pom. This is required in order to
  generate a Cygwin-compatible path.

- The `build_rpm.sh` execution is disabled by default (set to
  build phase none). It is enabled by a profile that is activated
  on unix or on Windows if `-Drpmbuild` is passed to the maven
  command-line.

The `-Drpmbuild` flag is required on Windows to execute the rpm
build because there are a number of requirements that must be met
in order for the rpm build tob be successful:

- Docker must be installed and working

- Cygwin must be installed with rpm packages

- Cygwin's /bin directory must be on the Windows PATH in order to
  build from a cmd or powershell terminal. This is not required
  if maven is executed from a Cygwin shell.